Add periodic db_events run to bot
A new file, db_events.py, was implemented under the botcommands module to manage database events. In the main bot script, a periodic task has been added that calls the 'run_db_events' function every 10 minutes. This parallel task logic is achieved in a new definition of the 'main' function in which both the bot.start and the periodic task are scheduled to run concurrently.