Add top Hacker News to morning report command
The morning report command of the bot now includes the top posts from Hacker News. This was achieved by refactoring the get_top_hacker_news() function to return a string with the top articles, which is then added to the existing message component. In addition, the no_school notification message has been relocated in the code.