Add plain text option to Bible bot command
Refined the 'get_esv_text' function in the Bible bot command to provide an option for outputting plain text. Also, the default test passage was updated from a placeholder to 'John 3:16'. More changes were included in the morse code bot command but the major update is the addition of the plain text option in the Bible bot command.